Least squares, changes to numpify
I wrote the function to do a least squares fit to a sum of general functions. It works as expected. Its doc must be written yet.
Some changes to numpify: it now returns by default column vectors, because we want to work with columns. It was a big pain to understand how the hell numpy works with matrices, columns and rows